+++ src/browser-utility.js
@@ -24,6 +24,14 @@ export function constantObjectValue(constant) {
 }
 
 /**
+ * @param {AbstractConstant} constant
+ * @returns {*}
+ */
+export function constantObjectValueOf(constant) {
+  return constant.valueOf();
+}
+
+/**
  * @param {AbstractBuilder} builder
  * @returns {{}}
  */